refactor(frontend): move logic of state about image to component
